pyproject.toml
[project]
name = "ai-morning-brief"
version = "0.1.0"
description = "Daily AI intelligence briefing from trending social posts"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.12.0",
    "arxiv>=2.1.0",
    "discord-webhook>=1.3.0",
    "trafilatura>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "responses>=0.24.0",
]
